Added TraderDetailsViewModelSpec.
