+++ app/controllers/traderDetails/CheckYourAnswersController.scala
@@ -22,7 +22,7 @@ import models.LocalReferenceNumber
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import viewModels.TraderDetailsViewModel
+import viewModels.traderDetails.TraderDetailsViewModel
 import views.html.traderDetails.CheckYourAnswersView
 
 class CheckYourAnswersController @Inject() (
