+++ app/controllers/traderDetails/holderOfTransit/CheckYourAnswersController.scala
@@ -22,7 +22,7 @@ import models.{LocalReferenceNumber, NormalMode}
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import viewModels.HolderOfTransitViewModel
+import viewModels.traderDetails.HolderOfTransitViewModel
 import views.html.traderDetails.holderOfTransit.CheckYourAnswersView
 
 class CheckYourAnswersController @Inject() (
