+++ app/viewModels/traderDetails/RepresentativeViewModel.scala
@@ -14,39 +14,29 @@
  * limitations under the License.
  */
 
-package viewModels
+package viewModels.traderDetails
 
 import models.{Mode, UserAnswers}
 import play.api.i18n.Messages
-import utils.HolderOfTransitCheckYourAnswersHelper
+import utils.RepresentativeCheckYourAnswersHelper
 import viewModels.sections.Section
 
-class HolderOfTransitViewModel {
+class RepresentativeViewModel {
 
   def apply(userAnswers: UserAnswers, mode: Mode)(implicit messages: Messages): Seq[Section] = {
-    val helper = new HolderOfTransitCheckYourAnswersHelper(userAnswers, mode)
+    val helper = new RepresentativeCheckYourAnswersHelper(userAnswers, mode)
 
-    val holderOfTransitSection = Section(
-      sectionTitle = messages("traderDetails.holderOfTransit.checkYourAnswers.transitHolder"),
+    val representativeSection = Section(
+      sectionTitle = messages("traderDetails.representative.checkYourAnswers.representative"),
       rows = Seq(
-        helper.eoriYesNo,
-        helper.eori,
-        helper.tirIdentificationYesNo,
-        helper.tirIdentification,
-        helper.name,
-        helper.address
+        helper.actingRepresentative,
+        helper.representativeEori,
+        helper.representativeName,
+        helper.representativeCapacity,
+        helper.representativePhone
       ).flatten
     )
 
-    val additionalContactSection = Section(
-      sectionTitle = messages("traderDetails.holderOfTransit.checkYourAnswers.additionalContact"),
-      rows = Seq(
-        helper.addContact,
-        helper.contactName,
-        helper.contactTelephoneNumber
-      ).flatten
-    )
-
-    Seq(holderOfTransitSection, additionalContactSection)
+    Seq(representativeSection)
   }
 }
